backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
databases
alembic
psycopg2-binary
python-multipart
Pillow
pytesseract
pyzbar
torch
sentence-transformers
faiss-cpu
pulp
requests
python-dotenv
httpx
